# internal/api/api.swagger.yaml
openapi: 3.0.3
info:
  title: Logiflow API
  description: API для логистической информационной системы
  version: 0.2.0
  contact:
    name: anxi0uz
servers:
  - url: http://localhost:3001
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        status:
          type: integer
          nullable: false
        data:
          type: object
          nullable: true
        success:
          type: boolean
          nullable: false
        requestID:
          type: string
          nullable: false
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        requestID:
          type: string

    # ─── Auth ─────────────────────────────────────────────────────────────────
...
    VehicleCreate:
      type: object
      required: [plateNumber]
      properties:
        plateNumber:
          type: string
        brand:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        year:
          type: integer
          nullable: true
        capacityKg:
          type: number
          nullable: true
        capacityM3:
          type: number
          nullable: true
    VehicleUpdate:
      type: object
      properties:
        brand:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        year:
          type: integer
          nullable: true
        capacityKg:
          type: number
          nullable: true
        capacityM3:
          type: number
          nullable: true
        status:
          type: string
          enum: [available, in_transit, maintenance]
          nullable: true

    # ─── Warehouse ────────────────────────────────────────────────────────────
...
paths:
...
  /vehicles:
    get:
      operationId: listVehicles
      summary: Список транспортных средств
      tags: [Vehicles]
      responses:
        "200":
          description: Список ТС
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: Не авторизован

    post:
      operationId: createVehicle
      summary: Добавить ТС
      tags: [Vehicles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleCreate"
      responses:
        "200":
          description: ТС создано
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /vehicles/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string

    get:
      operationId: getVehicle
      summary: Получить ТС по slug
      tags: [Vehicles]
      responses:
        "200":
          description: Данные ТС
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "404":
          description: Не найдено

    put:
      operationId: updateVehicle
      summary: Обновить ТС
      tags: [Vehicles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleUpdate"
      responses:
        "200":
          description: ТС обновлено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "404":
          description: Не найдено

    delete:
      operationId: deleteVehicle
      summary: Удалить ТС
      tags: [Vehicles]
      responses:
        "200":
          description: ТС удалено
        "404":
          description: Не найдено

  # ─── Warehouses ─────────────────────────────────────────────────────────────
...